requirements.txt
streamlit
requests
beautifulsoup4
dnspython
duckduckgo-search
google-genai
python-dotenv
fpdf